ES2015+ refactor
TODO: There's a bunch of "internal" public/exported/accessible functions that are not part of the original public API that need to be encapsulated again.
The tests need to get at them still, and I really like the simplicity with `export * from 'module'`. I don't want to have to register forwarding functions in the index like before.